Warning: Trying to access array offset on value of type bool in

+++ src/Suppliers/Reliable/Reliable.php
@@ -75,19 +75,22 @@ class getReliableActive
     
     public function addNipToReliable(string $nip):void
     {
-        $stmt = $this->PDO->prepare("SELECT * FROM weryfikacja_formalna where guid = (Select zgloszeni_wiarygodni.guid_wf from zgloszeni_wiarygodni where NIP=?)");
-        $stmt->execute(array($nip));
-        $data =  $stmt->fetch();
-        $stmt = $this->PDO->prepare("INSERT INTO lista_wiarygodnych (nip, guid, nazwa,data_dodania) VALUES (?,?,?,?)");
+        if(strlen($nip) == 10)
+        {
+            $stmt = $this->PDO->prepare("SELECT * FROM weryfikacja_formalna where guid = (Select zgloszeni_wiarygodni.guid_wf from zgloszeni_wiarygodni where NIP=?)");
+            $stmt->execute(array($nip));
+            $data =  $stmt->fetch();
+            $stmt = $this->PDO->prepare("INSERT INTO lista_wiarygodnych (nip, guid, nazwa,data_dodania) VALUES (?,?,?,?)");
             $stmt->execute(
                 array(
                     $data['nip'],
                     $data['guid'],
                     $data['nazwa'],
-                   date('Y-m-d H:i'),
+                    date('Y-m-d H:i'),
                 )
             );
-        $this->PDO->prepare("DELETE FROM zgloszeni_wiarygodni where nip=?")->execute(array($data['nip']));
+            $this->PDO->prepare("DELETE FROM zgloszeni_wiarygodni where nip=?")->execute(array($data['nip']));
+        }
 
 
     }
